oae_serial: remove use of actions and start commands
actions will be folded into commands, since editing commands is the same
as actions. The distinction between actions and commands is not clear,
and implementation-wise the difference between actions and commands is
nearly the same. Accordingly, removal of actions should simplify
extentions to the oae serial protocol.

there is some behavior related to stop commands that needs to be
solidified with the removal of actions (and later the start command)