cliffbar/serial: add reset command
the reset command is available as a command in the serial host; upon
reset, the device usb D+ and D- pins are pulled low to simulate a
physical usb disconnect. Additionally, the stm32's reset pin is pulled
low to reset the mcu and all registers.

for debugging purposes the device is reset once upon initial serial
connection such that the device is in a known state when connected via
usb